+++ src/client/toodee/canvas/circle.ts
@@ -1,7 +1,7 @@
 import { CircleGeometry } from '../geometry/circle_geometry'
 import { Shape } from '../object/shape'
 
-import { applyAppearance } from './canvas'
+import { applyAppearance } from './util'
 
 export function renderCircle(ctx: CanvasRenderingContext2D, shape: Shape<CircleGeometry>): void {
