+++ src/client/toodee/canvas/util.ts
